+++ src/app/_components/Header/UserProfileLinks.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import Profile from './Profile';
 import { dummyUserInfo } from '@/constants/dummy';
-import { AlarmSVG, CommentSVG, SearchSVG } from '../../../public/icons/svg';
+import { AlarmSVG, CommentSVG, SearchSVG } from '../../../../public/icons/svg';
 
 const UserProfileLinks = () => {
   const { alarmCount, commentCount } = dummyUserInfo;
