+++ src/app/class/create/_components/ClassPrice.tsx
@@ -1,0 +1,16 @@
+const ClassPrice = () => {
+  return (
+    <ul className="flex flex-col ">
+      <li className="flex">
+        <h2 className="w-1/4">총 클래스 횟수</h2>
+        <div>20회</div>
+      </li>
+      <li className="flex">
+        <h2 className="w-1/4">1회 최대 수강생</h2>
+        <div>20회</div>
+      </li>
+    </ul>
+  );
+};
+
+export default ClassPrice;
